reusable code for page size
