+++ apps/server/src/routers/index.ts
@@ -62,12 +62,6 @@ export const appRouter = {
       throw new Error("DB Not OK");
     }
   }),
-  privateData: protectedProcedure.handler(({ context }) => {
-    return {
-      message: "This is private",
-      user: context.session?.user,
-    };
-  }),
   auth: authRouter,
   categories: categoriesRouter,
   merchants: merchantsRouter,
