+++ apps/web/src/components/dashboard/merchant-stats.tsx
@@ -1,6 +1,7 @@
 import { useNavigate } from "@tanstack/react-router";
 import { StoreIcon } from "lucide-react";
 import { cn } from "@/lib/utils";
+import { getPreferredPageSize } from "@/utils/page-size";
 import type { DashboardMerchantStats } from "../../../../server/src/routers";
 import { CurrencyAmount } from "../ui/currency-amount";
 
@@ -32,7 +33,11 @@ export function MerchantStats({
   const handleMerchantClick = (merchantId: string) => {
     navigate({
       to: "/transactions",
-      search: { merchant: merchantId, page: 1 },
+      search: {
+        merchant: merchantId,
+        page: 1,
+        pageSize: getPreferredPageSize(),
+      },
     });
   };
